requirements.txt
pywin32
pillow
requests
